DIABLO-874 Add test coverage, remove deprecated views and tasks
