+++ config/test.py
@@ -49,6 +49,8 @@ INDEX_HTML = f'{BASE_DIR}/tests/static/test-index.html'
 
 LOGGING_LOCATION = 'STDOUT'
 
+SKIP_SIS_REFRESH_FOR_TESTING = True
+
 SQLALCHEMY_DATABASE_URI = 'postgresql://diablo:diablo@localhost:5432/pazuzu_test'
 
 TESTING = True
