+++ fixtures/kaltura/schedule_event.json
@@ -1,0 +1,5 @@
+{
+  "id": 666,
+  "name": "A Kaltura schedule event",
+  "templateEntryId": 1234567
+}
